Skip functions for which the current user does not have execute rights.
When SQLancer is triggered with non-superusers, it may not have
EXECUTE rights to all functions. This isn't the normal use-case however,
but it is a trivial fix to additionally accommodate this use-case.

with this change, existing use-cases (when superuser is used), should
work exactly as before, since superuser always has access to all
functions.